Add a description
This is a small application I created with HTML5, Javascript, jQuery and Materialize in order to display vehicle information from the REST API on https://www.vegvesen.no.
It features some custom Javascript coding to validate input with a Regex as well as dynamic display of the table displaying the information.